[TestCVEchecker] test npm
